Blaze: move the "Learn more" link into the page description
It sat next to the Promote button, where it read as a second button and
was easy to miss. It now sits at the end of the description that already
explains what Blaze does, where people are more likely to find it.

The description and the link are shared by the Blaze setup page, so it
shows up there too.
